fixed checks with accordance to qulice
