+++ gamification/src/main/java/juja/microservices/gamification/user/MongoUserRepository.java
@@ -1,21 +1,16 @@
 /**
  * Copyright (c) 2016, juja.com.ua
  * All rights reserved.
- *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- *
  * * Redistributions of source code must retain the above copyright notice, this
  * list of conditions and the following disclaimer.
- *
  * * Redistributions in binary form must reproduce the above copyright notice,
  * this list of conditions and the following disclaimer in the documentation
  * and/or other materials provided with the distribution.
- *
  * * Neither the name of microservices nor the names of its
  * contributors may be used to endorse or promote products derived from
  * this software without specific prior written permission.
- *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -32,14 +27,16 @@
 package juja.microservices.gamification.user;
 
 import java.util.List;
+import org.springframework.data.mongodb.repository.MongoRepository;
 
 /**
- * User custom repository interface.
+ * User repository interface.
  * @author Sergii Lisnychyi (ljore@ukr.net)
  * @version $Id$
  * @since 1.0
  */
-public interface UserRepo {
+public interface MongoUserRepository
+    extends MongoRepository<CommonUser, String>, UserRepository {
 
     /**
      * Create user.
@@ -49,7 +46,7 @@ public interface UserRepo {
     String createUser(String username);
 
     /**
-     * Get user by Uuid.
+     * Get user by uuid.
      * @param uuid Uuid
      * @return User
      */
